Crate m_captcha

mCaptcha is a proof of work based Denaial-of-Service attack protection system. This is is a server library that you can embed in your services to protect your servers.

A commercial managed solution is in the works but I'd much rather prefer folks host their own instances as it will make the more decentralized and free.

In mCaptcha, defense is adjusted in discrete levels that depend on the ammount of traffic that a service is experiencing. So users of this library are requested to benchmark their target machines before configuring their mCaptcha component.

Terminology:

Difficulty(Factor): Minimum ammount of work that a client must do to make a valid request.

Defense: A datatype that various visitor-difficulty mappigns

Visitor: Smallest unit of traffic, usually a single request. The more you have, the busier your service is. Determines mCaptcha defense defense

Visitor threshold: The threshold at which MCaptcha will adjust defense defense

Example:

use m_captcha::{LevelBuilder, cache::HashCache, DefenseBuilder, message::Visitor, MCaptchaBuilder};
// traits from actix needs to be in scope for starting actor
use actix::prelude::*;

#[actix_rt::main]
async fn main() -> std::io::Result<()> {
    // configure defense
    let defense = DefenseBuilder::default()
        // add as many defense as you see fit
        .add_level(
            LevelBuilder::default()
                // visitor_threshold is the threshold/limit at which
                // mCaptcha will adjust difficulty defense
                // it is advisable to set small values for the first
                // defense visitor_threshold and difficulty_factor
                // as this will be the work that clients will be
                // computing when there's no load
                .visitor_threshold(50)
                .difficulty_factor(500)
                .unwrap()
                .build()
                .unwrap(),
        )
        .unwrap()
        .add_level(
            LevelBuilder::default()
                .visitor_threshold(5000)
                .difficulty_factor(50000)
                .unwrap()
                .build()
                .unwrap(),
        )
        .unwrap()
        .build()
        .unwrap();

    //let cache = HashCache::default().start();

    // create and start MCaptcha actor
    let mcaptcha = MCaptchaBuilder::default()
        .defense(defense)
        // leaky bucket algorithm's emission interval
        .duration(30)
     //   .cache(cache)
        .build()
        .unwrap()
        .start();

    // increment count when user visits protected routes
    mcaptcha.send(Visitor).await.unwrap();

    Ok(())
}

Re-exports

pub use crate::cache::hashcache::HashCache;

pub use defense::Defense;

pub use defense::DefenseBuilder;

pub use defense::LevelBuilder;

pub use mcaptcha::MCaptcha;

pub use mcaptcha::MCaptchaBuilder;

Modules

cache
message datatypes to interact with MCaptcha actor

defense
Defense datatypes

errors
Errors and Result module

master

mcaptcha
MCaptcha actor module that manages defense levels

message
message datatypes to interact with MCaptcha actor