Module m_captcha::cache::messages

Messages that can be sent to cache data structures implementing Save

Structs

CachePoW
Message to cache PoW difficulty factor and string

CachePoWBuilder
Builder for CachePoW.

CacheResult
Message to cache captcha result and the captcha key for which it was generated

CacheResultBuilder
Builder for CacheResult.

CachedPoWConfig

DeleteCaptchaResult
Message to delete cached capthca result when it expires

DeletePoW
Message to delete cached PoW difficulty factor and string when they expire

RetrivePoW
Message to retrive the the difficulty factor for the specified string from the cache

VerifyCaptchaResult
Message to verify captcha result against the stored captcha key